requirements.txt
py-cord
python-whois
ping3
dnspython
asyncio
psutil
scapy
requests